# Avoid large allocations of temporary buffers in BuferUtil