--- a/jetty-util/src/main/java/org/eclipse/jetty/util/BufferUtil.java
+++ b/jetty-util/src/main/java/org/eclipse/jetty/util/BufferUtil.java
@@ -553,10 +553,12 @@ public class BufferUtil
         }
         else
         {
-            byte[] bytes = new byte[TEMP_BUFFER_SIZE];
+            byte[] bytes = null;
             while (buffer.hasRemaining())
             {
                 int byteCountToWrite = Math.min(buffer.remaining(), TEMP_BUFFER_SIZE);
+                if (bytes == null)
+                    bytes = new byte[byteCountToWrite];
                 buffer.get(bytes, 0, byteCountToWrite);
                 out.write(bytes, 0, byteCountToWrite);
             }