@gregw 2020-07-15
Thanks for this.  I'm reviewing now.
Meanwhile, do you have a stack trace showing where writeTo is being used from?