@gregw 2020-07-15
@sbordet @lorban , The [HttpContentRangeWriter](https://github.com/eclipse/jetty.project/blob/fb4cfa9be6c45d36bb85809f96de5433087b09e3/jetty-server/src/main/java/org/eclipse/jetty/server/resource/HttpContentRangeWriter.java#L49) is preferring a direct buffer, but [ByteBufferRangeWriter](https://github.com/eclipse/jetty.project/blob/fb4cfa9be6c45d36bb85809f96de5433087b09e3/jetty-server/src/main/java/org/eclipse/jetty/server/resource/ByteBufferRangeWriter.java#L62) always does a writeTo, causing a copy to a temp buffer.  This is a very suboptimal approach!