@SerCeMan 2020-07-15
Hey, @gregw! Thank you for the review, the feedback is addressed in https://github.com/eclipse/jetty.project/pull/5045/commits/99d19267910a03b7a5fe72ecc95daeeed42c4144 as suggested. 

> Meanwhile, do you have a stack trace showing where writeTo is being used from?

You might have already found out, but it's called from
https://github.com/eclipse/jetty.project/blob/6d88f966e1bfe910f03ac1179ee84877c47d01e2/jetty-websocket/websocket-common/src/main/java/org/eclipse/jetty/websocket/common/message/SimpleBinaryMessage.java#L61